requirements.txt
requests
pycryptodome
smart_open[s3]
